[inductor] Use python constants in IndexPropagation
In the next PR I have the IR `ops.neg(ops.constant(0.0, torch.float32))`
which should be folded to `ops.constant(-0.0, torch.float32)` but it seems that
`sympy.Float(-0.0)` doesn't respect the sign of the zero and so we instead
get a positive zero constant.

Here, I work around this by doing the constant folding with python arithmetic
which does respect signed zeros.

ghstack-source-id: 4033a00fd54bcec7bafb5dea58d0b329d929de33
Pull Request resolved: https://github.com/pytorch/pytorch/pull/122031